+++ inst/tinytest/test_slice_time.R
@@ -1,7 +1,7 @@
 library(gdalcubes)
 v = cube_view(srs = "EPSG:4326", extent = list(left = 5, right = 10, bottom = 48, top = 53, 
                                                t0 = "2021-01-01", t1 = "2021-12-31"), dt = "P1D", 
-              dx = 0.01, dy = 0.01)
+              dx = 0.05, dy = 0.05)
 v
 
 gdalcubes:::.raster_cube_dummy(v, 1, 1.0) |>